# Optimize Puma::Request#default_server_port