@MSP-Greg 2025-09-06
> I'm guessing this is sometimes a symbol

`Symbol#start_with?` was added in [Ruby 2.7](https://msp-greg.github.io/ruby_2_7/Core/Symbol.html#start_with?-instance_method), with a `String` parameter.

> Introducing constants actually decreases performance because it adds an extra lookup to the bytecode... because it's a frozen string.

Makes sense.